Added fixes after code review
