+++ .gitignore
@@ -21,3 +21,10 @@
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
+
+# IntellijIdea files
+.idea/
+
+# Project folder
+target/
+!target/site/jacoco/
